Split off post-in-progress code into its own file
